items controller updated and checked
