+++ src/REST/Manager.php
@@ -34,6 +34,7 @@ class Manager extends Singleton {
 				'\EverAccounting\Rest\ExpensesController',
 				'\EverAccounting\Rest\PaymentsController',
 				'\EverAccounting\Rest\RevenuesController',
+				'\EverAccounting\Rest\ItemsController',
 				'\EverAccounting\Rest\CategoriesController',
 				'\EverAccounting\Rest\CurrenciesController',
 				'\EverAccounting\Rest\TransfersController',
